+++ Libft/libft.h
@@ -55,7 +55,7 @@ void					ft_putchar_fd(char c, int fd);
 void					ft_putstr_fd(char *s, int fd);
 void					ft_putendl_fd(char *s, int fd);
 void					ft_putnbr_fd(int n, int fd);
-int						ft_atoi(const char *s);
+long						ft_atoi(const char *s);
 char					*ft_strdup(const char *s1);
 char					*ft_strtrim_left(char const *s1, char const *set);
